melib/text: cleanup wcwidth module
- Remove unnecessary traits, types
- Use char instead of u32 when appropriate
- split wcswidth/grapheme_clusters tests

+++ meli/src/terminal/cells.rs
@@ -834,7 +834,7 @@ impl CellBuffer {
                     .set_bg(bg_color)
                     .set_attrs(attrs);
 
-                match wcwidth(u32::from(c)) {
+                match wcwidth(c) {
                     Some(0) | None => {
                         // Skip drawing zero width characters
                         self[(x, y)].empty = true;
